# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "SavdoSklad biznes boshqaruv tizimi API hujjatlari",
        "title": "SavdoSklad API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/districts": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Geography"
                ],
                "summary": "Get all districts by region",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Region ID",
                        "name": "regionId",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/entity.District"
                            }
                        }
                    }
                }
            }
        },
...
        "/markets": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Geography"
                ],
                "summary": "Get all markets by district",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "District ID",
                        "name": "districtId",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/entity.Market"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "entity.District": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "regionId": {
                    "type": "integer"
                },
                "regionName": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "entity.Market": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "districtId": {
                    "type": "integer"
                },
                "districtName": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}